requirements_azure.txt
# Azure Blob Downloader Requirements
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
